Locating our workspace
This addition uses the os and shutil modules to locate ourselves in the
workspace. We can now automate our file renaming and cover replacements
and create a directory to do the rest of the regexes. Also messed
around with the files in here a bit more to better replicate the work
situation.